gud: Centralize timing data
The timing data used in the EDID did not match the timing data used by
gud when querying modes, with the pixel clock being set in the EDID but
not in the mode object. This lead to the refresh rate of the preferred
mode being set to 13mHz. The blanking values in the EDID were also
incorrect, leading to a negative backporch.

Introduce a struct for timing information as well as default timings,
and use that to fill out both the mode object and EDID. The comment
about drm not handling zero-sized blanking also does not seem accurate,
so the new default timings use zero blanking.